✨ feat(event-bus): update imports to use new surf.api.paper package structure
- refactor imports in multiple test files to align with new package structure
- rename ChatListener.java and CommandExceptionTest.java for consistency
- update usage of SurfVisualizerApi to SurfPaperVisualizerApi in visualizer tests
