+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/command/SurfApiTestCommand.java
@@ -1,0 +1,32 @@
+package dev.slne.surf.api.paper.test.command;
+
+import dev.jorel.commandapi.CommandAPICommand;
+import dev.slne.surf.api.paper.test.command.subcommands.*;
+import dev.slne.surf.surfapi.bukkit.test.command.subcommands.*;
+
+public class SurfApiTestCommand extends CommandAPICommand {
+
+    public SurfApiTestCommand() {
+        super("surfapitest");
+
+        withPermission("surfapitest.use");
+
+        withSubcommands(
+                new PacketLoreTest("packetlore"),
+                new ScoreboardTest("scoreboard"),
+                new SmoothTimeSkip("smoothtimeskip"),
+                new PacketEntityTest("packetentity"),
+                new ReflectionTest("reflection"),
+                new PrefixConfigTest("prefixconfig"),
+                new CommandExceptionTest("commandexception"),
+                new MaxStacksizeTest("maxstacksize"),
+                new VisualizerTest("visualizer"),
+                new GlowingTest("glowing"),
+                new PaginationTest("pagination"),
+                new InventoryTest("inventory"),
+                new ToastTest("toast"),
+                new SuspendCommandExecutionTest("suspendCommandExecution"),
+                new SummonCommandTest("summoncommand")
+        );
+    }
+}
