+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/command/subcommands/CommandExceptionTest.java
@@ -1,7 +1,7 @@
 package dev.slne.surf.api.paper.test.command.subcommands;
 
 import dev.jorel.commandapi.CommandAPICommand;
-import dev.slne.surf.api.core.api.command.SurfCommandUtil;
+import dev.slne.surf.api.core.command.SurfCommandUtil;
 import net.kyori.adventure.text.Component;
 
 public class CommandExceptionTest extends CommandAPICommand {
