+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/command/subcommands/MaxStacksizeTest.java
@@ -3,7 +3,7 @@ package dev.slne.surf.api.paper.test.command.subcommands;
 import dev.jorel.commandapi.CommandAPICommand;
 import dev.jorel.commandapi.arguments.IntegerArgument;
 import dev.jorel.commandapi.arguments.ItemStackArgument;
-import dev.slne.surf.api.paper.api.nms.bridges.SurfBukkitNmsItemBridge;
+import dev.slne.surf.api.paper.nms.bridges.SurfPaperNmsItemBridge;
 import org.bukkit.inventory.ItemStack;
 
 public class MaxStacksizeTest extends CommandAPICommand {
@@ -17,11 +17,10 @@ public class MaxStacksizeTest extends CommandAPICommand {
             final ItemStack item = args.getUnchecked("item");
             final int maxStackSize = args.getUnchecked("maxStackSize");
 
-            SurfBukkitNmsItemBridge.getInstance()
-                .setDefaultMaxStackSize(item.getType().asItemType(), maxStackSize);
+            SurfPaperNmsItemBridge.Companion.setDefaultMaxStackSize(item.getType().asItemType(), maxStackSize);
 
             sender.sendMessage(
-                "Set max stack size of " + item.getType().name() + " to " + maxStackSize);
+                    "Set max stack size of " + item.getType().name() + " to " + maxStackSize);
         });
     }
 }
