+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/command/subcommands/SmoothTimeSkip.java
@@ -1,13 +1,14 @@
 package dev.slne.surf.api.paper.test.command.subcommands;
 
 import dev.jorel.commandapi.CommandAPICommand;
-import dev.slne.surf.api.paper.api.SurfApiBukkit;
-import dev.slne.surf.api.paper.api.time.SkipOperations;
-import java.util.concurrent.CompletableFuture;
+import dev.slne.surf.api.paper.SurfApiPaper;
+import dev.slne.surf.api.paper.time.SkipOperations;
 import kotlinx.coroutines.CoroutineStart;
 import kotlinx.coroutines.Dispatchers;
 import kotlinx.coroutines.GlobalScope;
 import kotlinx.coroutines.future.FutureKt;
+
+import java.util.concurrent.CompletableFuture;
 
 public class SmoothTimeSkip extends CommandAPICommand {
 
@@ -16,10 +17,10 @@ public class SmoothTimeSkip extends CommandAPICommand {
 
         executes((commandSender, commandArguments) -> {
             CompletableFuture<Object> future = FutureKt.future(GlobalScope.INSTANCE,
-                Dispatchers.getIO(),
-                CoroutineStart.DEFAULT,
-                (coroutineScope, continuation) -> SurfApiBukkit
-                    .skipTimeSmoothly(SkipOperations.NEXT_DAY, continuation));
+                    Dispatchers.getIO(),
+                    CoroutineStart.DEFAULT,
+                    (coroutineScope, continuation) -> SurfApiPaper.Companion
+                            .skipTimeSmoothly(SkipOperations.NEXT_DAY, continuation));
 
             future.thenAccept((result) -> {
                 commandSender.sendMessage("Time skipped smoothly");
