+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/command/subcommands/reflection/Reflection.java
@@ -1,13 +1,14 @@
 package dev.slne.surf.api.paper.test.command.subcommands.reflection;
 
-import dev.slne.surf.api.core.api.reflection.SurfReflection;
+
+import dev.slne.surf.api.core.reflection.SurfReflection;
 
 public class Reflection {
 
     public static final SpigotRestartCommandProxy RESTART_COMMAND;
 
     static {
-        final SurfReflection reflection = SurfReflection.getInstance();
+        final SurfReflection reflection = SurfReflection.Companion;
 
         RESTART_COMMAND = reflection.createProxy(SpigotRestartCommandProxy.class);
     }
