+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/command/subcommands/reflection/SpigotRestartCommandProxy.java
@@ -1,8 +1,8 @@
 package dev.slne.surf.api.paper.test.command.subcommands.reflection;
 
 
-import dev.slne.surf.api.core.api.reflection.Static;
-import dev.slne.surf.api.core.api.reflection.SurfProxy;
+import dev.slne.surf.api.core.reflection.Static;
+import dev.slne.surf.api.core.reflection.SurfProxy;
 
 @SurfProxy(qualifiedName = "org.spigotmc.RestartCommand")
 public interface SpigotRestartCommandProxy {
