+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/config/TestConfig.java
@@ -1,17 +1,18 @@
 package dev.slne.surf.api.paper.test.config;
 
-import java.util.LinkedList;
-import java.util.List;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.spongepowered.configurate.objectmapping.ConfigSerializable;
+
+import java.util.LinkedList;
+import java.util.List;
 
 @ConfigSerializable
 public class TestConfig {
 
     public LinkedList<String> testLinkedList = new LinkedList<>(List.of("test1", "test2", "test3"));
     public LinkedList<TestObject> testObjectLinkedList = new LinkedList<>(
-        List.of(new TestObject(1, "test1"), new TestObject(2, "test2"),
-            new TestObject(3, "test3")));
+            List.of(new TestObject(1, "test1"), new TestObject(2, "test2"),
+                    new TestObject(3, "test3")));
 
     @ConfigSerializable
     public static class TestObject {
