+++ surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/api/paper/test/example/MainExample.java
@@ -1,0 +1,5 @@
+package dev.slne.surf.api.paper.test.example;
+
+public class MainExample {
+
+}
