--- surf-api-paper/surf-api-paper-plugin-test/src/main/java/dev/slne/surf/surfapi/bukkit/example/MainExample.java
@@ -1,5 +1,0 @@
-package dev.slne.surf.api.paper.example;
-
-public class MainExample {
-
-}
