+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/dialog/subcommands/PaginatedDialogTest.kt
@@ -7,18 +7,18 @@ import com.github.shynixn.mccoroutine.folia.scope
 import dev.jorel.commandapi.CommandAPICommand
 import dev.jorel.commandapi.kotlindsl.playerExecutor
 import dev.jorel.commandapi.kotlindsl.subcommand
-import dev.slne.surf.surfapi.bukkit.api.dialog.base
-import dev.slne.surf.surfapi.bukkit.api.dialog.builder.actionButton
-import dev.slne.surf.surfapi.bukkit.api.dialog.clearDialogs
-import dev.slne.surf.surfapi.bukkit.api.dialog.dialog
-import dev.slne.surf.surfapi.bukkit.api.dialog.pagination.paginatedDialog
-import dev.slne.surf.surfapi.bukkit.api.dialog.query.DialogQuery
-import dev.slne.surf.surfapi.bukkit.api.dialog.query.PageResult
-import dev.slne.surf.surfapi.bukkit.api.dialog.query.PageState
-import dev.slne.surf.surfapi.bukkit.api.dialog.type
-import dev.slne.surf.surfapi.bukkit.api.nms.NmsUseWithCaution
+import dev.slne.surf.api.core.messages.adventure.buildText
+import dev.slne.surf.api.paper.dialog.base
+import dev.slne.surf.api.paper.dialog.builder.actionButton
+import dev.slne.surf.api.paper.dialog.clearDialogs
+import dev.slne.surf.api.paper.dialog.dialog
+import dev.slne.surf.api.paper.dialog.pagination.paginatedDialog
+import dev.slne.surf.api.paper.dialog.query.DialogQuery
+import dev.slne.surf.api.paper.dialog.query.PageResult
+import dev.slne.surf.api.paper.dialog.query.PageState
+import dev.slne.surf.api.paper.dialog.type
+import dev.slne.surf.api.paper.nms.NmsUseWithCaution
 import dev.slne.surf.surfapi.bukkit.test.plugin
-import dev.slne.surf.surfapi.core.api.messages.adventure.buildText
 import io.papermc.paper.dialog.Dialog
 import net.kyori.adventure.text.Component
