+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/subcommands/SummonCommandTest.kt
@@ -6,9 +6,9 @@ import dev.jorel.commandapi.CommandAPICommand
 import dev.jorel.commandapi.kotlindsl.entityTypeArgument
 import dev.jorel.commandapi.kotlindsl.getValue
 import dev.jorel.commandapi.kotlindsl.playerExecutor
-import dev.slne.surf.surfapi.bukkit.api.command.args.adventureCompoundBinaryTagArgument
-import dev.slne.surf.surfapi.bukkit.api.nms.NmsUseWithCaution
-import dev.slne.surf.surfapi.bukkit.api.nms.bridges.entityBridge
+import dev.slne.surf.api.paper.command.args.adventureCompoundBinaryTagArgument
+import dev.slne.surf.api.paper.nms.NmsUseWithCaution
+import dev.slne.surf.api.paper.nms.bridges.SurfPaperNmsEntityBridge
 import net.kyori.adventure.nbt.CompoundBinaryTag
 import org.bukkit.entity.EntityType
 
@@ -21,7 +21,7 @@ class SummonCommandTest(name: String) : CommandAPICommand(name) {
             val type: EntityType by args
             val nbt: CompoundBinaryTag by args
 
-            entityBridge.createEntityByNbt(sender.world, type, sender.location, nbt)
+            SurfPaperNmsEntityBridge.createEntityByNbt(sender.world, type, sender.location, nbt)
         }
     }
 }
