+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/subcommands/inventory/TestInventoryView.kt
@@ -1,18 +1,18 @@
 package dev.slne.surf.surfapi.bukkit.test.command.subcommands.inventory
 
-import dev.slne.surf.surfapi.bukkit.api.builder.buildItem
-import dev.slne.surf.surfapi.bukkit.api.builder.displayName
-import dev.slne.surf.surfapi.bukkit.api.dialog.noticeDialog
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.dsl.slot
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.dsl.withItem
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.titleBuilder
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.view.onClose
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.view.onFirstRender
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.view.state.get
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.view.state.increment
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.view.state.mutableState
-import dev.slne.surf.surfapi.bukkit.api.inventory.framework.view.surfView
-import dev.slne.surf.surfapi.core.api.messages.adventure.text
+import dev.slne.surf.api.core.messages.adventure.text
+import dev.slne.surf.api.paper.builder.buildItem
+import dev.slne.surf.api.paper.builder.displayName
+import dev.slne.surf.api.paper.dialog.noticeDialog
+import dev.slne.surf.api.paper.inventory.framework.dsl.slot
+import dev.slne.surf.api.paper.inventory.framework.dsl.withItem
+import dev.slne.surf.api.paper.inventory.framework.titleBuilder
+import dev.slne.surf.api.paper.inventory.framework.view.onClose
+import dev.slne.surf.api.paper.inventory.framework.view.onFirstRender
+import dev.slne.surf.api.paper.inventory.framework.view.state.get
+import dev.slne.surf.api.paper.inventory.framework.view.state.increment
+import dev.slne.surf.api.paper.inventory.framework.view.state.mutableState
+import dev.slne.surf.api.paper.inventory.framework.view.surfView
 import me.devnatan.inventoryframework.View
 import me.devnatan.inventoryframework.ViewConfigBuilder
 import me.devnatan.inventoryframework.ViewType
