+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/subcommands/suspendexecution/SuspendCommandExecutionCancellationExceptionTest.kt
@@ -1,8 +1,8 @@
 package dev.slne.surf.surfapi.bukkit.test.command.subcommands.suspendexecution
 
 import dev.jorel.commandapi.CommandAPICommand
-import dev.slne.surf.surfapi.bukkit.api.command.executors.anyExecutorSuspend
-import dev.slne.surf.surfapi.core.api.messages.adventure.sendText
+import dev.slne.surf.api.core.messages.adventure.sendText
+import dev.slne.surf.api.paper.command.executors.anyExecutorSuspend
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.delay
 import kotlin.time.Duration.Companion.seconds
