+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/subcommands/suspendexecution/SuspendCommandExecutionDelayWithSuspendArgTest.kt
@@ -3,10 +3,10 @@ package dev.slne.surf.surfapi.bukkit.test.command.subcommands.suspendexecution
 import dev.jorel.commandapi.CommandAPICommand
 import dev.jorel.commandapi.arguments.StringArgument
 import dev.jorel.commandapi.kotlindsl.arguments
-import dev.slne.surf.surfapi.bukkit.api.command.args.SuspendCustomArgument
-import dev.slne.surf.surfapi.bukkit.api.command.executors.anyExecutorSuspend
-import dev.slne.surf.surfapi.core.api.command.args.awaiting
-import dev.slne.surf.surfapi.core.api.messages.adventure.sendText
+import dev.slne.surf.api.core.command.args.awaiting
+import dev.slne.surf.api.core.messages.adventure.sendText
+import dev.slne.surf.api.paper.command.args.SuspendCustomArgument
+import dev.slne.surf.api.paper.command.executors.anyExecutorSuspend
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
 import org.bukkit.Bukkit
