+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/subcommands/suspendexecution/SuspendCommandExecutionSyntaxExceptionTest.kt
@@ -2,12 +2,12 @@ package dev.slne.surf.surfapi.bukkit.test.command.subcommands.suspendexecution
 
 import dev.jorel.commandapi.CommandAPI
 import dev.jorel.commandapi.CommandAPICommand
-import dev.slne.surf.surfapi.bukkit.api.command.executors.anyExecutorSuspend
-import dev.slne.surf.surfapi.core.api.messages.adventure.sendText
+import dev.slne.surf.api.core.messages.adventure.sendText
+import dev.slne.surf.api.paper.command.executors.anyExecutorSuspend
 import kotlinx.coroutines.delay
 import kotlin.time.Duration.Companion.seconds
 
-class SuspendCommandExecutionSyntaxExceptionTest(name: String): CommandAPICommand(name) {
+class SuspendCommandExecutionSyntaxExceptionTest(name: String) : CommandAPICommand(name) {
     init {
         anyExecutorSuspend { sender, arguments ->
             sender.sendText {
