+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/config/ModernTestConfig.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.surfapi.bukkit.test.config
 
+import dev.slne.surf.api.core.config.SpongeYmlConfigClass
 import dev.slne.surf.surfapi.bukkit.test.plugin
-import dev.slne.surf.surfapi.core.api.config.SpongeYmlConfigClass
 import org.spongepowered.configurate.objectmapping.ConfigSerializable
 
 @ConfigSerializable
