+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/config/MyPluginConfig.kt
@@ -1,8 +1,8 @@
 package dev.slne.surf.surfapi.bukkit.test.config
 
+import dev.slne.surf.api.core.config.SpongeYmlConfigClass
+import dev.slne.surf.api.core.config.migration.ConfigMigration
 import dev.slne.surf.surfapi.bukkit.test.plugin
-import dev.slne.surf.surfapi.core.api.config.SpongeYmlConfigClass
-import dev.slne.surf.surfapi.core.api.config.migration.ConfigMigration
 import org.spongepowered.configurate.ConfigurationNode
 import org.spongepowered.configurate.objectmapping.ConfigSerializable
