+++ surf-api-paper/surf-api-paper-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/hook/condition/EnabledCondition.kt
@@ -1,8 +1,8 @@
 package dev.slne.surf.surfapi.bukkit.test.hook.condition
 
+import dev.slne.surf.api.shared.api.component.condition.ComponentCondition
+import dev.slne.surf.api.shared.api.component.condition.ComponentConditionContext
 import dev.slne.surf.surfapi.bukkit.test.config.ModernTestConfig
-import dev.slne.surf.surfapi.shared.api.component.condition.ComponentCondition
-import dev.slne.surf.surfapi.shared.api.component.condition.ComponentConditionContext
 
 class EnabledCondition : ComponentCondition {
     override suspend fun evaluate(context: ComponentConditionContext): Boolean {
